requirements.txt
# Core dependencies
# mediapipe==0.10.18
# latest verwsion mediapipe is 0.10.21
opencv-python==4.12.0.88
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0

# GazeTracking dependencies
dlib>=19.24.0
scipy>=1.10.0
gaze-tracking>=0.0.1

# Logging and configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Performance monitoring
psutil>=5.9.0

# Data handling
pandas>=2.0.0
h5py>=3.8.0  # For future HDF5 support

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# API dependencies (for future integration)
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0